requirements.txt
# Hybrid Voice Chatbot - Python Dependencies
# Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Session==0.5.0
Flask-Babel==4.0.0

# Database
SQLAlchemy==2.0.23
pymysql==1.1.0
pymongo==4.6.0

# AIML Engine
python-aiml==0.9.3

# Voice Processing
gTTS==2.5.0
# pyttsx3==2.90  # Disabled for cloud deployment (requires system audio drivers)
# SpeechRecognition==3.10.0  # Disabled for cloud deployment
# pyaudio==0.2.14  # Commented for cloud deployment (requires system dependencies)

# NLP & Text Processing
nltk==3.8.1
textblob==0.17.1

# Security & Auth
Werkzeug==3.0.1
python-dotenv==1.0.0
bcrypt==4.1.2
bleach==6.1.0
PyJWT==2.8.0

# Image Processing
Pillow==12.0.0

# Utilities
requests==2.31.0

# Production Server
gunicorn==21.2.0
waitress==2.1.2

# Testing (install separately for development)
pytest==8.0.0
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0
locust==2.20.0

# Code Quality (install separately for development)
mypy==1.8.0
black==24.1.1
flake8==7.0.0
pylint==3.0.3

# Monitoring
prometheus-client==0.19.0

# Redis (for caching and rate limiting - optional)
redis==5.0.1

# Environment management
python-decouple==3.8
